pyproject.toml
[tool.poetry]
authors = ["William Sampaio <56176702+WilliamSampaio@users.noreply.github.com>"]
description = ""
name = "planodeleituraanual"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
beautifulsoup4 = "^4.12.3"
lxml = "^5.2.2"
numpy = "^1.26.4"
odfpy = "^1.4.1"
pandas = "^2.2.2"
plotly = "^5.22.0"
pyinstaller = "^6.7.0"
python = ">=3.12,<3.13"
python-dotenv = "^1.0.1"
streamlit = "^1.35.0"
tinydb = "^4.8.0"

[tool.poetry.group.dev.dependencies]
blue = "^0.9.1"
isort = "^5.13.2"
taskipy = "^1.12.2"

[tool.isort]
line_length = 79
profile = "black"

[tool.taskipy.tasks]
freeze = "poetry export --output=requirements.txt"
lint = "blue . && isort ."
run = "task lint && streamlit run main.py"
# build = "task lint && pyinstaller -D -F --hidden-import 'pandas' --hidden-import 'numpy' --hidden-import 'streamlit' -n planodeleituraanual -c main.py"
build = "task lint && pyinstaller planodeleituraanual.spec"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
